pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "git-diff-analyzer"
version = "0.1.0"
description = "AI-powered git commit analysis tool that generates PDF docs and Snowflake DML scripts"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.0",
    "gitpython>=3.1.40",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "fpdf2>=2.7.9",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "rich>=13.7.0",
    "django>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "pytest-django>=4.8.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-pyyaml>=6.0.12",
    "boto3-stubs[bedrock-runtime]>=1.34.0",
]

[project.scripts]
git-diff-analyzer = "git_diff_analyzer.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
DJANGO_SETTINGS_MODULE = "analyzer_web.settings"

[tool.coverage.run]
source = ["git_diff_analyzer"]
